+++ packages/Webkul/Admin/src/Config/system.php
@@ -1898,6 +1898,9 @@ return [
                     ], [
                         'title' => '6',
                         'value' => 6,
+                    ], [
+                        'title' => '7',
+                        'value' => 7,
                     ],
                 ],
             ],
@@ -2010,6 +2013,9 @@ return [
                     ], [
                         'title' => '6',
                         'value' => 6,
+                    ], [
+                        'title' => '7',
+                        'value' => 7,
                     ],
                 ],
             ],
@@ -2107,6 +2113,9 @@ return [
                     ], [
                         'title' => '6',
                         'value' => 6,
+                    ], [
+                        'title' => '7',
+                        'value' => 7,
                     ],
                 ],
             ],
@@ -2183,6 +2192,9 @@ return [
                     ], [
                         'title' => '6',
                         'value' => 6,
+                    ], [
+                        'title' => '7',
+                        'value' => 7,
                     ],
                 ],
             ],
@@ -2293,6 +2305,9 @@ return [
                     ], [
                         'title' => '6',
                         'value' => 6,
+                    ], [
+                        'title' => '7',
+                        'value' => 7,
                     ],
                 ],
             ],
@@ -2300,7 +2315,7 @@ return [
     ], [
         'key'    => 'sales.payment_methods.payu',
         'name'   => 'admin::app.configuration.index.sales.payment-methods.payu',
-        'info'   => 'admin::app.configuration.index.sales.payment-methods.payment-method-info',
+        'info'   => 'admin::app.configuration.index.sales.payment-methods.payu-info',
         'sort'   => 6,
         'fields' => [
             [
@@ -2376,6 +2391,107 @@ return [
                     ], [
                         'title' => '6',
                         'value' => 6,
+                    ], [
+                        'title' => '7',
+                        'value' => 7,
+                    ],
+                ],
+            ],
+        ],
+    ], [
+        'key'    => 'sales.payment_methods.stripe',
+        'name'   => 'admin::app.configuration.index.sales.payment-methods.stripe',
+        'info'   => 'admin::app.configuration.index.sales.payment-methods.stripe-info',
+        'sort'   => 7,
+        'fields' => [
+            [
+                'name'          => 'active',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.status',
+                'type'          => 'boolean',
+                'channel_based' => true,
+                'locale_based'  => false,
+            ], [
+                'name'          => 'title',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.title',
+                'type'          => 'text',
+                'validation'    => 'required_if:active,1',
+                'channel_based' => true,
+                'locale_based'  => true,
+            ], [
+                'name'          => 'description',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.description',
+                'type'          => 'textarea',
+                'channel_based' => true,
+                'locale_based'  => true,
+            ], [
+                'name'          => 'image',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.logo',
+                'type'          => 'image',
+                'info'          => 'admin::app.configuration.index.sales.payment-methods.logo-information',
+                'channel_based' => false,
+                'locale_based'  => false,
+                'validation'    => 'mimes:bmp,jpeg,jpg,png,webp',
+            ], [
+                'name'          => 'debug',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.debug',
+                'type'          => 'boolean',
+                'channel_based' => true,
+                'locale_based'  => false,
+            ], [
+                'name'          => 'api_key',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.api_key',
+                'info'          => 'admin::app.configuration.index.sales.payment-methods.applicable-production',
+                'type'          => 'password',
+                'channel_based' => false,
+                'locale_based'  => false,
+            ], [
+                'name'          => 'api_publishable_key',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.api_publishable_key',
+                'info'          => 'admin::app.configuration.index.sales.payment-methods.applicable-production',
+                'type'          => 'password',
+                'channel_based' => false,
+                'locale_based'  => false,
+            ], [
+                'name'          => 'api_test_key',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.api_test_key',
+                'info'          => 'admin::app.configuration.index.sales.payment-methods.applicable-sandbox',
+                'type'          => 'password',
+                'channel_based' => false,
+                'locale_based'  => false,
+            ], [
+                'name'          => 'api_test_publishable_key',
+                'title'         => 'admin::app.configuration.index.sales.payment-methods.api_test_publishable_key',
+                'info'          => 'admin::app.configuration.index.sales.payment-methods.applicable-sandbox',
+                'type'          => 'password',
+                'channel_based' => false,
+                'locale_based'  => false,
+            ], [
+                'name'    => 'sort',
+                'title'   => 'admin::app.configuration.index.sales.payment-methods.sort-order',
+                'type'    => 'select',
+                'default' => 7,
+                'options' => [
+                    [
+                        'title' => '1',
+                        'value' => 1,
+                    ], [
+                        'title' => '2',
+                        'value' => 2,
+                    ], [
+                        'title' => '3',
+                        'value' => 3,
+                    ], [
+                        'title' => '4',
+                        'value' => 4,
+                    ], [
+                        'title' => '5',
+                        'value' => 5,
+                    ], [
+                        'title' => '6',
+                        'value' => 6,
+                    ], [
+                        'title' => '7',
+                        'value' => 7,
                     ],
                 ],
             ],