+++ packages/Webkul/Sitemap/src/Providers/SitemapServiceProvider.php
@@ -14,7 +14,5 @@ class SitemapServiceProvider extends ServiceProvider
     public function boot()
     {
         $this->loadMigrationsFrom(__DIR__.'/../Database/Migrations');
-
-        $this->app->register(ModuleServiceProvider::class);
     }
 }