+++ packages/Webkul/Stripe/src/Config/payment-methods.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+    'stripe' => [
+        'code'          => 'stripe',
+        'title'         => 'Stripe',
+        'description'   => 'Stripe Payments',
+        'class'         => 'Webkul\Stripe\Payment\Stripe',
+        'debug'         => true,
+        'active'        => false,
+    ],
+];